# Improve performances of BigInteger.numberOfDecimalDigits().